:recycle: 엘라스틱서치 호스트 암호화, tenantId 암호화 추가
